Add sessionHashAlg and move more pcr functions to pcr.go
