+++ tpm2tools/keys.go
@@ -176,7 +176,6 @@ func (k *Key) Seal(pcrs []int, sensitive []byte) (*proto.SealedBytes, error) {
 	if err != nil {
 		return nil, fmt.Errorf("could not get pcr session auth: %v", err)
 	}
-
 	sb, err := sealHelper(k.rw, k.Handle(), auth, sensitive)
 	if err != nil {
 		return nil, err
@@ -265,60 +264,4 @@ func (k *Key) Reseal(pcrs map[int][]byte, in *proto.SealedBytes) (*proto.SealedB
 	sb.Hash = in.Hash
 	sb.Srk = in.Srk
 	return sb, nil
-}
-
-type tpmsPCRSelection struct {
-	Hash tpm2.Algorithm
-	Size byte
-	PCRs tpmutil.RawBytes
-}
-
-type sessionSummary struct {
-	OldDigest      tpmutil.RawBytes
-	CmdIDPolicyPCR uint32
-	NumPcrSels     uint32
-	Sel            tpmsPCRSelection
-	PcrDigest      tpmutil.RawBytes
-}
-
-func getPCRSessionAuth(rw io.ReadWriter, pcrs []int, pcrHash tpm2.Algorithm) ([]byte, error) {
-	handle, err := createPCRSession(rw, pcrs, pcrHash)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get digest: %v", err)
-	}
-	defer tpm2.FlushContext(rw, handle)
-
-	digest, err := tpm2.PolicyGetDigest(rw, handle)
-	if err != nil {
-		return nil, fmt.Errorf("could not get digest from session: %v", err)
-	}
-
-	return digest, nil
-}
-
-func createPCRSession(rw io.ReadWriter, pcrs []int, pcrHash tpm2.Algorithm) (tpmutil.Handle, error) {
-	nonceIn := make([]byte, 16)
-	/* This session assumes the bus is trusted.  */
-	handle, _, err := tpm2.StartAuthSession(
-		rw,
-		tpm2.HandleNull,
-		tpm2.HandleNull,
-		nonceIn,
-		/*secret=*/ nil,
-		tpm2.SessionPolicy,
-		tpm2.AlgNull,
-		tpm2.AlgSHA256)
-	if err != nil {
-		return tpm2.HandleNull, fmt.Errorf("failed to start auth session: %v", err)
-	}
-
-	sel := tpm2.PCRSelection{
-		Hash: pcrHash,
-		PCRs: pcrs,
-	}
-	if err = tpm2.PolicyPCR(rw, handle, nil, sel); err != nil {
-		return tpm2.HandleNull, fmt.Errorf("auth step PolicyPCR failed: %v", err)
-	}
-
-	return handle, nil
 }
